Move log utils into its own package.
